+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -124,6 +124,7 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   impSatisfyMNativeScripts,
   impSatisfySignature,
   shelleyGenRegTxCert,
+  shelleyGenUnRegTxCert,
 
   -- * Logging
   Doc,
@@ -510,6 +511,8 @@ class
 
   genRegTxCert :: HasCallStack => Credential 'Staking -> ImpTestM era (TxCert era)
 
+  genUnRegTxCert :: HasCallStack => Credential 'Staking -> ImpTestM era (TxCert era)
+
 impSatisfySignature ::
   KeyHash 'Witness ->
   Set.Set (KeyHash 'Witness) ->
@@ -804,6 +807,7 @@ instance
   expectTxSuccess = impShelleyExpectTxSuccess
   modifyImpInitProtVer = shelleyModifyImpInitProtVer
   genRegTxCert = shelleyGenRegTxCert
+  genUnRegTxCert = shelleyGenUnRegTxCert
 
 -- | Figure out all the Byron Addresses that need witnesses as well as all of the
 -- KeyHashes for Shelley Key witnesses that are required.
@@ -1862,3 +1866,10 @@ shelleyGenRegTxCert ::
   Credential 'Staking ->
   ImpTestM era (TxCert era)
 shelleyGenRegTxCert = pure . RegTxCert
+
+shelleyGenUnRegTxCert ::
+  forall era.
+  ShelleyEraTxCert era =>
+  Credential 'Staking ->
+  ImpTestM era (TxCert era)
+shelleyGenUnRegTxCert = pure . UnRegTxCert